Renamed spec to test.
