+++ .rubocop.yml
@@ -1,6 +1,7 @@
+---
 AllCops:
   TargetRubyVersion: 2.5
 
 Metrics/BlockLength:
   Exclude:
-    - spec/**/*
+    - test/**/*
