+++ Rakefile
@@ -7,9 +7,9 @@ require 'rake/testtask'
 RuboCop::RakeTask.new
 
 Rake::TestTask.new(:test) do |t|
-  t.libs << 'spec'
+  t.libs << 'test'
   t.libs << 'lib'
-  t.test_files = FileList['spec/**/*_spec.rb']
+  t.test_files = FileList['test/**/*_test.rb']
 end
 
 task default: %i[rubocop test]
