+++ test/fixtures/valid_receipt.txt
